Re-home python_tree_root in health.rs and share interpreter_in_tree
Pure move ahead of #335. python_tree_root moves out of base_manifest.rs,
which #335 deletes wholesale, into health.rs where is_managed_python_tree
already needs it. Its inverse, the tests-only interpreter_in_tree helper,
is promoted into health.rs beside it and now also serves get_python_path,
replacing two cfg'd path pairs; the layouts stay verifiably in agreement
by sitting next to each other.
